New implimintations

+++ src/main/java/com/demo99/backend/services/Implementations/AvailabilityServiceImpl.java
@@ -43,6 +43,9 @@ public class AvailabilityServiceImpl {
     }
 
     public void deleteAvailability(Long id){
-        availabilityRepository.deleteById(id);
+        Optional<Availability> availabilityOptional = availabilityRepository.findById(id);
+        if (availabilityOptional.isPresent()){
+            availabilityRepository.deleteById(id);
+        }
     }
 }
